Convert HBase ttl param from Integer to long#1454

+++ janusgraph-hbase-parent/janusgraph-hbase-core/src/main/java/org/janusgraph/diskstorage/hbase/HBaseStoreManager.java
@@ -944,7 +944,7 @@ public class HBaseStoreManager extends DistributedStoreManager implements KeyCol
                             // Convert ttl from second (JanusGraph TTL) to milliseconds (HBase TTL)
                             // @see JanusGraphManagement#setTTL(JanusGraphSchemaType, Duration)
                             // HBase supports cell-level TTL for versions 0.98.6 and above.
-                            (putColumnWithTtl).setTTL(ttl * 1000);
+                            (putColumnWithTtl).setTTL((long)ttl * 1000);
                             // commands.getFirst() is the list of Puts for this rowkey. Add this
                             // Put column with TTL to the list.
                             commands.getFirst().add(putColumnWithTtl);
